requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
boto3==1.35.49
botocore==1.35.49
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
fastapi==0.115.3
filelock==3.16.1
firebase-admin==6.5.0
fonttools==4.54.1
fsspec==2024.10.0
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.67.0
grpcio-status==1.67.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
Jinja2==3.1.4
jmespath==1.0.1
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
numpy==2.1.2
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
pillow==11.0.0
proto-plus==1.25.0
protobuf==5.28.3
psutil==6.1.0
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
s3transfer==0.10.3
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
starlette==0.41.0
sympy==1.13.1
torch==2.5.0
torchvision==0.20.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.23
ultralytics-thop==2.0.9
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
